refactor: 자동차 이름 유효성 클래스 메소드로 분리해서 사용 되도록 수정
