+++ src/test/java/racingcar/RacingGameTest.java
@@ -1,11 +1,13 @@
 package racingcar;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import racingcar.controller.RacingGame;
+import racingcar.exception.CarNameException;
 
 public class RacingGameTest {
 
@@ -23,4 +25,11 @@ public class RacingGameTest {
     assertThat(racingGame.cars.size()).isEqualTo(3);
   }
 
+  @ParameterizedTest(name = "[RacingGame] validateName 테스트 - {index}")
+  @ValueSource(strings = {"", "abcdef", "abcdefg"})
+  void validateNameTest(String input) {
+    assertThatThrownBy(() -> {
+      racingGame.validateName(input);
+    }).isInstanceOf(CarNameException.class);
+  }
 }
